WIP REFACTOR Current version installs only non stan dependent part
Created a separate module for the error_propagation. Files paths and imports are altered to fit new setup.

This version was tested by locally install the package in fresh venv and run test_integration and test_unit
