+++ pseudobatch/error_propagation/stan/functions.stan
