+++ setup.cfg
@@ -13,11 +13,8 @@ packages = find:
 include_package_data = True
 requires_python = >=3.9
 install_requires =
-    arviz
-    cmdstanpy>=1.0.7
     numpy
     pandas
-    pydantic==2.4.2
 
 [options.package_data]
 * =
@@ -37,3 +34,7 @@ documentation =
     sphinx
     nbsphinx
     pydata-sphinx-theme
+error_propagation =
+    arviz
+    cmdstanpy>=1.0.7
+    pydantic==2.4.2
